Refactor AppWrapper to Elementary
